refactor(openclaw-config): split 1097-line module into focused sub-files (#233)
Move `packages/web/src/lib/openclaw-config.ts` into a directory with
focused files. Each new file has a single concern, and the upstream-bug
workarounds (openclaw#75534, tracked in #215) are now isolated in
`normalize.ts` so their eventual removal is a single-file delete.

  build.ts          regenerateOpenClawConfig (DB → openclaw.json)
  targeted.ts       sanitizeOpenClawConfig, updateIdentityLinks,
                    updateTelegramChannelConfig
  write.ts          writeConfigAtomic, readExistingConfig (EACCES retry),
                    pushConfigInBackground
  normalize.ts      configsAreEquivalentUpToOpenClawMetadata,
                    redactUnchangedEnvForApply, OPENCLAW_REDACTED_SENTINEL
                    — workarounds for openclaw#75534, removable per #215
  secrets-bundle.ts buildSecretsBundle helper + central home for the
                    Pattern A/B/C secret-handling matrix
  paths.ts          CONFIG_PATH constant (avoids write↔normalize cycle)
  index.ts          re-exports the public surface

The import path `@/lib/openclaw-config` is unchanged — `index.ts`
re-exports the four public functions, so all 10 production importers
and 14 test files keep working without edits.

Behavior is byte-for-byte identical:
- Trailing-newline format (trimEnd + "\n") preserved so SHA256 hashes
  align with OpenClaw's writeConfigFile output (#193 dedup).
- Byte-equal early return + normalize-compare order unchanged.
- Sentinel redaction semantics unchanged.
- Atomic write + EACCES retry semantics unchanged.

Verification: tsc --noEmit clean (exit 0). vitest run: 3407 passed,
12 skipped, 3 todo, 0 failed across 259 test files.
