+++ packages/web/src/lib/openclaw-config/build.ts
@@ -1,12 +1,5 @@
-import { writeFileSync, readFileSync, existsSync, mkdirSync, renameSync } from "fs";
-import { dirname } from "path";
-import { assertNoPlaintextSecrets } from "@/lib/openclaw-plaintext-scanner";
-import {
-  writeSecretsFile,
-  readSecretsFile,
-  secretRef,
-  type SecretsBundle,
-} from "@/lib/openclaw-secrets";
+import { readFileSync } from "fs";
+import { writeSecretsFile, readSecretsFile, secretRef } from "@/lib/openclaw-secrets";
 import { PROVIDERS, type ProviderName } from "@/lib/providers";
 import { getDefaultModel } from "@/lib/provider-models";
 import { eq, ne } from "drizzle-orm";
@@ -23,86 +16,12 @@ import type { AgentPluginConfig } from "@/db/schema";
 import { TOOL_CAPABLE_OLLAMA_CLOUD_MODELS, OLLAMA_CLOUD_COST } from "@/lib/ollama-cloud-models";
 import { getOpenClawWorkspacePath } from "@/lib/workspace";
 import { migrateExistingSmithers } from "@/lib/migrate-onboarding";
+import type { SecretsBundle } from "@/lib/openclaw-secrets";
 
-const CONFIG_PATH = process.env.OPENCLAW_CONFIG_PATH || "/openclaw-config/openclaw.json";
-
-/**
- * Remove stale Pinchy plugins from the allow list that have no matching entry.
- * OpenClaw validates config schemas for allowed plugins — if a plugin is in
- * `allow` but has no `entries` config, OpenClaw rejects the config and refuses
- * to start. This can happen when an older config volume is reused with a fresh
- * DB (e.g. after deleting pgdata). Runs before setup is complete (no DB needed).
- */
-export function sanitizeOpenClawConfig(): boolean {
-  if (!existsSync(CONFIG_PATH)) return false;
-
-  const raw = readFileSync(CONFIG_PATH, "utf-8");
-  const config = JSON.parse(raw);
-  const plugins = config.plugins as Record<string, unknown> | undefined;
-  if (!plugins) return false;
-
-  const allow = plugins.allow as string[] | undefined;
-  const entries = (plugins.entries ?? {}) as Record<string, unknown>;
-  if (!allow) return false;
-
-  const cleaned = allow.filter((p) => !p.startsWith("pinchy-") || p in entries);
-
-  if (cleaned.length === allow.length) return false;
-
-  plugins.allow = cleaned;
-  writeConfigAtomic(JSON.stringify(config, null, 2).trimEnd() + "\n");
-  return true;
-}
-
-/** Atomic write: tmp file + rename to prevent OpenClaw reading a truncated config */
-function writeConfigAtomic(content: string) {
-  const dir = dirname(CONFIG_PATH);
-  if (!existsSync(dir)) {
-    mkdirSync(dir, { recursive: true });
-  }
-  // Defense-in-depth: never let a plaintext secret land in openclaw.json.
-  assertNoPlaintextSecrets(JSON.parse(content));
-  const tmpPath = CONFIG_PATH + ".tmp";
-  writeFileSync(tmpPath, content, { encoding: "utf-8", mode: 0o644 });
-  renameSync(tmpPath, CONFIG_PATH);
-}
-
-function readExistingConfig(): Record<string, unknown> {
-  // Retry briefly on EACCES. OpenClaw rewrites openclaw.json as root:0600 on
-  // every internal SIGUSR1 restart; start-openclaw.sh's 3s chmod loop opens
-  // it back up to 0666, but Pinchy (uid 999) can hit a small window where
-  // the file is unreadable. Without retry, readFileSync throws → catch
-  // returns {} → targeted writes (updateTelegramChannelConfig etc.) would
-  // produce a config WITHOUT the gateway block, and OpenClaw's next start
-  // refuses with "Gateway start blocked: existing config is missing
-  // gateway.mode". 5 × 100ms covers two chmod-loop ticks worst case.
-  for (let attempt = 0; attempt < 5; attempt++) {
-    try {
-      return JSON.parse(readFileSync(CONFIG_PATH, "utf-8"));
-    } catch (err) {
-      const code = (err as NodeJS.ErrnoException)?.code;
-      if (code !== "EACCES") {
-        // ENOENT (file not yet written) is a normal cold-start case; other
-        // errors (parse failures, etc.) are bugs we can't paper over here.
-        return {};
-      }
-      if (attempt === 4) {
-        console.warn(
-          "[openclaw-config] readExistingConfig: persistent EACCES on",
-          CONFIG_PATH,
-          "— returning empty (callers must guard against partial writes)"
-        );
-        return {};
-      }
-      // Synchronous busy-wait. Async would change all caller signatures.
-      const start = Date.now();
-      while (Date.now() - start < 100) {
-        // spin
-      }
-    }
-  }
-  return {};
-}
+import { CONFIG_PATH } from "./paths";
+import { configsAreEquivalentUpToOpenClawMetadata } from "./normalize";
+import { writeConfigAtomic, readExistingConfig, pushConfigInBackground } from "./write";
+import { buildSecretsBundle } from "./secrets-bundle";
 
 function deepMerge(
   target: Record<string, unknown>,
@@ -755,12 +674,12 @@ export async function regenerateOpenClawConfig() {
 
   // Always write secrets.json — tmpfs is wiped on container restart, secrets.json
   // must be present for OpenClaw to resolve SecretRef pointers (provider API keys etc.).
-  const secretsBundle: SecretsBundle = {
+  const secretsBundle = buildSecretsBundle({
     gateway: gatewaySecret,
     providers: providerSecrets,
     integrations: integrationSecrets,
     env: envSecrets,
-  };
+  });
   writeSecretsFile(secretsBundle);
 
   // Only write if content actually changed — prevents unnecessary OpenClaw restarts.
@@ -803,295 +722,4 @@ export async function regenerateOpenClawConfig() {
   // interactive save flows (Odoo permissions Save & Restart, where the
   // UI waits for "All changes saved").
   pushConfigInBackground(newContent);
-}
-
-/**
- * Compare two openclaw.json strings for semantic equivalence, ignoring
- * fields that OpenClaw stamps onto the file independently of any user
- * change. Used to short-circuit redundant config writes / config.apply
- * RPCs that would otherwise trigger spurious gateway restarts via
- * openclaw#75534. See call site for full rationale and removal tracking.
- *
- * Currently normalized: `meta.lastTouchedAt` (a write-time timestamp).
- * Add other OpenClaw-managed metadata fields here if they ever surface.
- */
-function configsAreEquivalentUpToOpenClawMetadata(a: string, b: string): boolean {
-  try {
-    const pa = JSON.parse(a) as Record<string, unknown>;
-    const pb = JSON.parse(b) as Record<string, unknown>;
-    const stripMeta = (cfg: Record<string, unknown>) => {
-      const meta = cfg.meta as Record<string, unknown> | undefined;
-      if (!meta) return;
-      delete meta.lastTouchedAt;
-      // If meta becomes empty after stripping, remove it entirely so an
-      // absent-meta config (cold start) compares equal to a meta-with-only-
-      // lastTouchedAt config (post-OpenClaw-stamp).
-      if (Object.keys(meta).length === 0) delete cfg.meta;
-    };
-    stripMeta(pa);
-    stripMeta(pb);
-    return JSON.stringify(pa) === JSON.stringify(pb);
-  } catch {
-    return false;
-  }
-}
-
-// OpenClaw's redacted-value sentinel. When OpenClaw sees this string in
-// `env.<KEY>` (or any other registered redactable path) inside an incoming
-// config.apply payload, it restores the corresponding value from
-// snapshot.config before running the change-paths diff. We rely on this
-// to avoid openclaw#75534's spurious env.* restart trigger: Pinchy sends
-// the sentinel for env keys whose template hasn't changed, OpenClaw
-// substitutes the resolved snapshot value, diffConfigPaths finds no env
-// diff, no restart.
-//
-// Defined in OpenClaw's `runtime-schema-*.js` as
-//   const REDACTED_SENTINEL = "__OPENCLAW_REDACTED__";
-// Stable since 2026.4.x. Removable when openclaw#75534 lands; tracked in #215.
-const OPENCLAW_REDACTED_SENTINEL = "__OPENCLAW_REDACTED__";
-
-/**
- * Replace `env.<KEY>` values with OpenClaw's redacted sentinel for keys
- * whose template form is unchanged from `existingContent`. New env keys
- * (not present in existing) keep their template form — they represent a
- * legitimate configuration change and the restart they trigger is valid.
- *
- * On cold start (no existing file), returns the input unchanged.
- *
- * This is the workaround for openclaw#75534: see comment on the sentinel
- * constant above. Removable when upstream lands the writer-level fix.
- */
-function redactUnchangedEnvForApply(newContent: string): string {
-  let existingContent: string;
-  try {
-    existingContent = readFileSync(CONFIG_PATH, "utf-8");
-  } catch {
-    return newContent;
-  }
-  try {
-    const newCfg = JSON.parse(newContent) as Record<string, unknown>;
-    const existingCfg = JSON.parse(existingContent) as Record<string, unknown>;
-    const newEnv = (newCfg.env as Record<string, string>) ?? {};
-    const existingEnv = (existingCfg.env as Record<string, string>) ?? {};
-    const redactedEnv: Record<string, string> = {};
-    for (const [key, val] of Object.entries(newEnv)) {
-      if (key in existingEnv && existingEnv[key] === val) {
-        redactedEnv[key] = OPENCLAW_REDACTED_SENTINEL;
-      } else {
-        redactedEnv[key] = val;
-      }
-    }
-    if (Object.keys(redactedEnv).length === 0 && !("env" in newCfg)) {
-      return newContent;
-    }
-    newCfg.env = redactedEnv;
-    return JSON.stringify(newCfg, null, 2);
-  } catch {
-    return newContent;
-  }
-}
-
-function pushConfigInBackground(newContent: string): void {
-  void (async () => {
-    let client;
-    try {
-      const { getOpenClawClient } = await import("@/server/openclaw-client");
-      client = getOpenClawClient();
-    } catch {
-      // No client — file write + inotify is the only path here.
-      return;
-    }
-
-    // Workaround for openclaw#75534: replace unchanged env values with
-    // OpenClaw's REDACTED sentinel before sending. Without this, every
-    // config.apply payload trips OpenClaw's resolved-vs-template diff for
-    // env.* paths and triggers a full gateway restart even when only a
-    // hot-reloadable path (agents.list, bindings) actually changed.
-    // Removable when openclaw#75534 lands; tracked in #215.
-    const payload = redactUnchangedEnvForApply(newContent);
-
-    // Brief retry across transient WS disconnects. Beyond ~3.5 s the WS is
-    // probably down due to the cold-start cascade, and inotify will catch
-    // up; no point keeping a background coroutine alive longer.
-    const backoffsMs = [100, 250, 500, 1000, 2000];
-    for (let i = 0; i < backoffsMs.length; i++) {
-      try {
-        const current = (await client.config.get()) as { hash: string };
-        await client.config.apply(payload, current.hash, {
-          note: "pinchy: regenerateOpenClawConfig",
-        });
-        return;
-      } catch (err) {
-        if (i === backoffsMs.length - 1) {
-          const message = err instanceof Error ? err.message : String(err);
-          console.warn(
-            "[openclaw-config] background config.apply failed; relying on inotify:",
-            message
-          );
-          return;
-        }
-        await new Promise((resolve) => setTimeout(resolve, backoffsMs[i]));
-      }
-    }
-  })();
-}
-
-// ── Targeted config updates ───────────────────────────────────────────────
-
-/**
- * Update only session.identityLinks in the config file.
- *
- * Unlike regenerateOpenClawConfig(), this reads the existing config and only
- * modifies session.identityLinks. All other fields (agents.defaults, env,
- * plugins, channels, meta, etc.) are preserved byte-for-byte. This avoids
- * unnecessary diffs that trigger hot-reloads breaking Telegram polling
- * (openclaw#47458).
- *
- * OpenClaw treats identityLinks changes as "dynamic reads" — no channel
- * restart, no hot-reload, just updated in memory.
- */
-export function updateIdentityLinks(identityLinks: Record<string, string[]>): void {
-  const existing = readExistingConfig();
-
-  // Same safety as updateTelegramChannelConfig — see comment there.
-  const existingGateway = existing.gateway as Record<string, unknown> | undefined;
-  if (!existingGateway?.mode) {
-    throw new Error(
-      "[openclaw-config] updateIdentityLinks: existing config has no gateway.mode " +
-        "(likely EACCES race on /openclaw-config/openclaw.json). Retry the request."
-    );
-  }
-
-  const session = (existing.session as Record<string, unknown>) || {};
-  const updatedSession = {
-    ...session,
-    identityLinks,
-  };
-
-  const updated = { ...existing, session: updatedSession };
-
-  // Only write if content actually changed. Format matches OpenClaw's
-  // writeConfigFile output (trimEnd + "\n") so SHA256 hashes line up
-  // for the reload-subsystem dedup. See call site of regenerateOpenClawConfig
-  // for the full rationale.
-  const newContent = JSON.stringify(updated, null, 2).trimEnd() + "\n";
-  try {
-    const current = readFileSync(CONFIG_PATH, "utf-8");
-    if (current === newContent) return;
-  } catch {
-    // File doesn't exist — write it
-  }
-
-  writeConfigAtomic(newContent);
-}
-
-/**
- * Update a single Telegram account in the config (add or remove).
- *
- * Uses OpenClaw's multi-account format: channels.telegram.accounts.<accountId>.
- * Preserves all other accounts, bindings, and OpenClaw-enriched fields.
- *
- * Pass `account: null` to remove an account. When the last account is removed,
- * the entire telegram channel config is removed.
- *
- * Used by bot connect/disconnect to avoid full config regeneration, which
- * overwrites OpenClaw-enriched fields (agents.defaults.*) and triggers
- * hot-reloads that break Telegram polling (openclaw#47458).
- */
-export function updateTelegramChannelConfig(
-  accountId: string | null,
-  account: { botToken: string } | null,
-  identityLinks: Record<string, string[]> | null
-): void {
-  const existing = readExistingConfig();
-
-  // Safety: refuse to write a config that would clobber the gateway block.
-  // If readExistingConfig returned an empty/partial object (EACCES race or
-  // corrupted file), modifying channels/bindings on top of it and writing
-  // back would produce a config without gateway.mode — OpenClaw refuses
-  // to start with "Gateway start blocked: existing config is missing
-  // gateway.mode" and falls into a restart loop. Throwing here lets the
-  // calling API route return 503 to the user instead of silently dropping
-  // the channel update.
-  const existingGateway = existing.gateway as Record<string, unknown> | undefined;
-  if (!existingGateway?.mode) {
-    throw new Error(
-      "[openclaw-config] updateTelegramChannelConfig: existing config has no gateway.mode " +
-        "(likely EACCES race on /openclaw-config/openclaw.json). Retry the request."
-    );
-  }
-
-  if (accountId && account) {
-    const existingTelegram =
-      ((existing.channels as Record<string, unknown>)?.telegram as Record<string, unknown>) || {};
-    const existingAccounts = (existingTelegram.accounts as Record<string, unknown>) || {};
-
-    existingAccounts[accountId] = { botToken: account.botToken };
-
-    existing.channels = {
-      ...((existing.channels as Record<string, unknown>) || {}),
-      telegram: {
-        ...existingTelegram,
-        dmPolicy: "pairing",
-        accounts: existingAccounts,
-      },
-    };
-
-    // Update bindings: add this account's binding, preserve others
-    const existingBindings =
-      (existing.bindings as Array<{ agentId: string; match: Record<string, string> }>) || [];
-    const otherBindings = existingBindings.filter((b) => b.match?.accountId !== accountId);
-    existing.bindings = [
-      ...otherBindings,
-      { agentId: accountId, match: { channel: "telegram", accountId } },
-    ];
-  } else if (accountId && !account) {
-    const existingTelegram =
-      ((existing.channels as Record<string, unknown>)?.telegram as Record<string, unknown>) || {};
-    const existingAccounts = (existingTelegram.accounts as Record<string, unknown>) || {};
-
-    delete existingAccounts[accountId];
-
-    if (Object.keys(existingAccounts).length === 0) {
-      // Last account removed — remove entire telegram config
-      const channels = (existing.channels as Record<string, unknown>) || {};
-      delete channels.telegram;
-      existing.channels = Object.keys(channels).length > 0 ? channels : undefined;
-      existing.bindings = undefined;
-    } else {
-      existing.channels = {
-        ...((existing.channels as Record<string, unknown>) || {}),
-        telegram: { ...existingTelegram, accounts: existingAccounts },
-      };
-      // Remove this account's binding
-      const existingBindings =
-        (existing.bindings as Array<{ agentId: string; match: Record<string, string> }>) || [];
-      existing.bindings = existingBindings.filter((b) => b.match?.accountId !== accountId);
-    }
-  } else {
-    // No accountId — remove ALL telegram config (used by remove-all)
-    const channels = (existing.channels as Record<string, unknown>) || {};
-    delete channels.telegram;
-    existing.channels = Object.keys(channels).length > 0 ? channels : undefined;
-    existing.bindings = undefined;
-  }
-
-  const session = (existing.session as Record<string, unknown>) || {};
-  existing.session = {
-    ...session,
-    dmScope: "per-peer",
-    // null = "don't touch existing identityLinks" (used by bot connect/disconnect).
-    // Non-null = overwrite with provided value (used by link/unlink).
-    ...(identityLinks !== null && { identityLinks }),
-  };
-
-  const newContent = JSON.stringify(existing, null, 2).trimEnd() + "\n";
-  try {
-    const current = readFileSync(CONFIG_PATH, "utf-8");
-    if (current === newContent) return;
-  } catch {
-    // File doesn't exist
-  }
-
-  writeConfigAtomic(newContent);
 }
